Optimize project card links for mobile
Changed behavior of anchors linking to projects so that user has
to click on project title to open the project link in a new tab.
Previous behaviour had the entire overlay text wrapped in anchor
so clicking to reveal the overlay on mobile automatically redirected
to the project link before user had a chance to read the text.

Changed appearance of project title to make it more apparent this
is a clickable link.